refactor: corrected single table design approach in dynamoose
