+++ packages/user-articles/src/repositories/DynamoDbRepository/DynamoDbRepository.ts
@@ -2,105 +2,19 @@ import * as dynamoose from 'dynamoose';
 import { config } from '../../config';
 import { ModelType } from 'dynamoose/dist/General';
 import { AnyItem } from 'dynamoose/dist/Item';
+import { UserModel } from '../../models/user';
+import { UserSessionModel } from '../../models/userSession';
 
-export const dbUserSchema = new dynamoose.Schema(
-  {
-    pk: {
-      hashKey: true,
-      map: 'Login',
-      type: String,
-    },
-    sk: {
-      default: 'ROOT',
-      rangeKey: true,
-      type: String,
-    },
-    Password: String,
-    UserId: String,
-  },
-  {
-    timestamps: {
-      createdAt: {
-        CreatedAt: {
-          type: {
-            value: String,
-            settings: {
-              storage: 'iso',
-            },
-          },
-        },
-      },
-      updatedAt: {
-        UpdatedAt: {
-          type: {
-            value: String,
-            settings: {
-              storage: 'iso',
-            },
-          },
-        },
-      },
-    },
-  },
-);
-
-export const dbUserSessionSchema = new dynamoose.Schema(
-  {
-    pk: {
-      hashKey: true,
-      map: 'Login',
-      type: String,
-    },
-    sk: {
-      rangeKey: true,
-      type: String,
-    },
-    Token: {
-      index: true,
-      type: String,
-    },
-    UserId: String,
-  },
-  {
-    timestamps: {
-      createdAt: {
-        CreatedAt: {
-          type: {
-            value: String,
-            settings: {
-              storage: 'iso',
-            },
-          },
-        },
-      },
-      updatedAt: {
-        UpdatedAt: {
-          type: {
-            value: String,
-            settings: {
-              storage: 'iso',
-            },
-          },
-        },
-      },
-    },
-  },
-);
-
-export const UserModel = dynamoose.model(
-  'User',
-  [dbUserSchema, dbUserSessionSchema],
-  {
-    create: config.environment === 'local',
-    update: config.environment === 'local',
-    waitForActive: config.environment === 'local',
-  },
-);
+new dynamoose.Table('UserArticle', [UserModel, UserSessionModel]);
 
 export class DynamoDbRepository {
   model: ModelType<AnyItem>;
 
   constructor(model: ModelType<AnyItem>) {
     this.model = model;
+
+    if (config.dbEndpointUrl) {
+      dynamoose.aws.ddb.local(config.dbEndpointUrl);
+    }
   }
 }
